Cleaned up optimizer code
